Extract in-memory SQS implementation to its own module
